FieldAuditApp with QR thing and BottomSlider done with Example POST Api done and 2nd API done also

+++ app/src/main/java/com/chalo/fieldauditapp/model/Fine.kt
@@ -2,5 +2,5 @@ package com.chalo.fieldauditapp.model
 
 data class Fine(
     val amount: Int,
-    val fine_levied_at: Int
+    val fine_levied_at: Long
 )
